Update README.md
# Sound And Fury OPerationS (safops)

This GitHub organization is managed by the servants of the Creator and will provide information about the Creator's great storm and war against senseless death and those who support it.

## About the Creator 

The Creator that we work for is the very one who created life and the universe that we all live it in. 

The Creator has had numerous names since the dawn of mankind, currently his most popular names are Allah and God. Regardless of what you call Him, or not, the Creator the only one who can ever hear your  prayers or read your mind. The Creator does not require worship but doesn't mind if you do.

He gets beside himself though, as you have felt, when doves (peaceful people) are made to cry.

We ask that anytime you refer to the Creator in writing you always use a capital "C," and when discussing the Creator, you always clarify that it is the Creator with the "big C."  There are entities out there who are trying to confuse others as to who we work for.


# Welcome